Allow show manager to configure show seats/tickets
Add page showing auditorium layout along with input to enter price and
whether it is reserved for vip or not for each row.

The configuration including number of rows, columns, etc are stored in
an auditorium configuration file and the layout is being generated on
hitting the page itself for now.

Allow to include scripts in the main admin layout.

Also for admin shows list page is giving unauthorized action. This is
coming because it is hitting the role middleware first and since for
shows both show managers and audi sec have same routes so it is going
with the role of show_manager when audi sec is logged in. So change the
priority of middleware, so that first role based redirect middleware
would hit first then the corresponding controller would hit and before
hitting that the role middleware would be hit.