Fix: seats not getting reserved for vip
In the controller method, we were checking for if the vip field is set,
and if it was then we were setting the reserved flag to 1 else 0 but a
hidden field for those rows which are not marked for reserved send a
value of off, so in the request body the 'vip' field is set so it was
marking it as reserved.

Instead of this, just check the value of 'vip' to 'on', if yes then 1
else 0.